tidy up local notification
