+++ lib/src/notification/daily_reflection_reminder.dart
@@ -1,0 +1,98 @@
+import 'dart:async';
+import 'dart:io';
+
+import 'package:ekayanaarama/src/routes/middle_path.dart';
+import 'package:ekayanaarama/src/routes/route_name.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_local_notifications/flutter_local_notifications.dart';
+import 'package:flutter_timezone/flutter_timezone.dart';
+import 'package:get/get.dart';
+import 'package:timezone/data/latest_all.dart' as tz;
+import 'package:timezone/timezone.dart' as tz;
+
+final FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
+    FlutterLocalNotificationsPlugin();
+
+Future<void> _configureLocalTimeZone() async {
+  if (kIsWeb || Platform.isLinux) {
+    return;
+  }
+  tz.initializeTimeZones();
+  final String timeZoneName = await FlutterTimezone.getLocalTimezone();
+  tz.setLocalLocation(tz.getLocation(timeZoneName));
+}
+
+tz.TZDateTime _nextInstanceOfTenAM() {
+   tz.TZDateTime now = tz.TZDateTime.now(tz.local);
+  // tz.TZDateTime scheduledDate =
+  //     tz.TZDateTime(tz.local, now.year, now.month, now.day, 10);
+  // if (scheduledDate.isBefore(now)) {
+  //   scheduledDate = scheduledDate.add(const Duration(days: 1));
+  // }
+  now = now.add(const Duration(seconds: 15));
+  return now;
+}
+
+@pragma('vm:entry-point')
+void _localNotificationTapBackground(NotificationResponse notificationResponse) {
+  // ignore: avoid_print
+  print('notification(${notificationResponse.id}) action tapped: '
+      '${notificationResponse.actionId} with'
+      ' payload: ${notificationResponse.payload}');
+  if (notificationResponse.input?.isNotEmpty ?? false) {
+    // ignore: avoid_print
+    print(
+        'notification action tapped with input: ${notificationResponse.input}');
+  }
+}
+
+Future<String?> _setupLocalPushNotification() async {
+  const AndroidInitializationSettings initializationSettingsAndroid = AndroidInitializationSettings('app_icon');
+  const DarwinInitializationSettings initializationSettingsDarwin = DarwinInitializationSettings();
+
+  const InitializationSettings initializationSettings = InitializationSettings(
+    android: initializationSettingsAndroid,
+    iOS: initializationSettingsDarwin,
+  );
+  await flutterLocalNotificationsPlugin.initialize(
+    initializationSettings,
+    onDidReceiveNotificationResponse: (NotificationResponse notificationResponse) {
+      Get.toNamed(RouteName.dailyReflection);
+    },
+    onDidReceiveBackgroundNotificationResponse: _localNotificationTapBackground,
+  );
+
+  final NotificationAppLaunchDetails? notificationAppLaunchDetails = await flutterLocalNotificationsPlugin.getNotificationAppLaunchDetails();
+
+  if (notificationAppLaunchDetails?.didNotificationLaunchApp ?? false) {
+    return RouteName.dailyReflection;
+  }  
+  return null;
+}
+
+Future<String?> setupDailyReflectionReminder() async {
+  await _configureLocalTimeZone();
+  final result = await _setupLocalPushNotification();
+  return result;
+}
+
+void setScheduleNotification() async {
+  await flutterLocalNotificationsPlugin.zonedSchedule(
+    0,
+    'Ekayana Arama Indonesia Buddhist Center',
+    'Renungan Hari Ini',
+    _nextInstanceOfTenAM(),
+    const NotificationDetails(
+      android: AndroidNotificationDetails(
+        'daily_reflection_id',
+        'daily_reflection_name',
+        channelDescription: 'daily notification description',
+      ),
+    ),
+    androidScheduleMode: AndroidScheduleMode.exactAllowWhileIdle,
+    uiLocalNotificationDateInterpretation:
+        UILocalNotificationDateInterpretation.absoluteTime,
+    matchDateTimeComponents: DateTimeComponents.time,
+  );
+}
