+++ lib/src/routes/middle_path.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+class MiddlePath {
+  static final GlobalKey<NavigatorState> navState = GlobalKey<NavigatorState>();
+}
